requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
mysqlclient
pydantic-settings
fastapi-pagination
fastapi-cache
slowapi
os
base64
pytest
mysql-connector-python
uuid
sqlalchemy-utils
smtplib
MIMEText
MIMEMultipart

